moved functions to functions.py
